# app/public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/groups": {
      "post": {
        "tags": [
          "Group API"
        ],
        "operationId": "App\\Controllers\\GroupController::create",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "example": "junior php group"
                  },
                  "min_students_num": {
                    "example": 3
                  },
                  "max_students_num": {
                    "example": 20
                  },
                  "min_skills_num": {
                    "example": 5
                  },
                  "max_skills_num": {
                    "example": 6
                  },
                  "max_useless_skill_students": {
                    "example": "1"
                  },
                  "skills": {
                    "example": "[2, 3, 4]"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/groups/{id}": {
      "get": {
        "tags": [
          "Group API"
        ],
        "operationId": "App\\Controllers\\GroupController::search",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The identifier.",
            "required": true,
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Group API"
        ],
        "operationId": "App\\Controllers\\GroupController::delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The identifier.",
            "required": true,
            "example": 1
          }
        ],
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Group API"
        ],
        "operationId": "App\\Controllers\\GroupController::update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The identifier.",
            "required": true,
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "example": "middle php group"
                  },
                  "min_students_num": {
                    "example": 3
                  },
                  "max_students_num": {
                    "example": 20
                  },
                  "min_skills_num": {
                    "example": 5
                  },
                  "max_skills_num": {
                    "example": 6
                  },
                  "max_useless_skill_students": {
                    "example": "1"
                  },
                  "enabled": {
                    "example": false
                  },
                  "skills": {
                    "example": "[3, 4, 5, 6]"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/groups/{id}/find-teacher": {
      "post": {
        "tags": [
          "Group API"
        ],
        "operationId": "App\\Controllers\\GroupController::findTeacher",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The identifier.",
            "required": true,
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "example": "middle php group"
                  },
                  "min_students_num": {
                    "example": 3
                  },
                  "max_students_num": {
                    "example": 20
                  },
                  "min_skills_num": {
                    "example": 5
                  },
                  "max_skills_num": {
                    "example": 6
                  },
                  "max_useless_skill_students": {
                    "example": "1"
                  },
                  "enabled": {
                    "example": false
                  },
                  "skills": {
                    "example": "[3, 4, 5, 6]"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/groups/{id}/change-teacher": {
      "patch": {
        "tags": [
          "Group API"
        ],
        "operationId": "App\\Controllers\\GroupController::changeTeacher",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The identifier.",
            "required": true,
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "example": "middle php group"
                  },
                  "min_students_num": {
                    "example": 3
                  },
                  "max_students_num": {
                    "example": 20
                  },
                  "min_skills_num": {
                    "example": 5
                  },
                  "max_skills_num": {
                    "example": 6
                  },
                  "max_useless_skill_students": {
                    "example": "1"
                  },
                  "enabled": {
                    "example": false
                  },
                  "skills": {
                    "example": "[3, 4, 5, 6]"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/groups/{id}/form-group": {
      "post": {
        "tags": [
          "Group API"
        ],
        "operationId": "App\\Controllers\\GroupController::formGroup",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The identifier.",
            "required": true,
            "example": 1
          }
        ],
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/requests/{id}": {
      "get": {
        "tags": [
          "QueueRequest API"
        ],
        "operationId": "App\\Controllers\\RequestController::getStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The identifier.",
            "required": true,
            "example": 1
          }
        ],
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}